changed from long to int

+++ src/main/java/main/IViewModel.java
@@ -6,6 +6,6 @@ public interface IViewModel
 
 	void startWorking();
 	void startResting();
-	void setSeconds(long seconds);
+	void setSeconds(int seconds);
 
 }
